[navigation] Add special handling for "&jqmNPopups=[0-9][0-9]*"
In addition to the special handling for "&ui-state=dialog", special handling for
"&jqmNPopups=[0-9][0-9]*" is also needed, in order to avoid attempts at loading
pages upon hashchange.

Currently this assumes that the presence of a dialog and that of one or more
popups is mutually exclusive. That is, either "&ui-state=dialog" is present, or
"&jqmNPopups=[0-9][0-9]*" is present, but both are not.